[rename] Filenames & includes: half --> float16_t
